+++ changelog.md
@@ -3,13 +3,14 @@
 ## 1.19.3 (TBD)
 
 - Added masteries for Lightwing, Magunet, Derail Gun, Alien Nailgun, Vladimir, and Maestro
-- Added 1 new synergy
+- Added 2 new synergies
 
 - Fixed potential null deref caused by Magunet being destroyed while holding debris in stasis
 - Fixed Alien Nailgun losing all DNA information when dropped and picked back up
 - Vladimir now has piercing and can hit multiple enemies simultaneously
 - Gorgun Eye can now target non-hostile enemies such as Keybullet Kin and Chance Kin
 - Changed Gorgun Eye's quality from B to A (extremely powerful, especially in single enemy rooms or against strong enemies)
+- Changed base damage of all of Breegull's eggs from 5 to 7
 
 ## 1.19.2 (2024-09-29)